refactor: move typecasting to client
