+++ alumnium/clients/native_client.py
@@ -1,8 +1,8 @@
-from ..server.agents.retriever_agent import Data
 from ..server.models import Model
 from ..server.session_manager import SessionManager
 from ..tools.base_tool import BaseTool
 from ..tools.tool_to_schema_converter import convert_tools_to_schemas
+from .typecasting import Data, loosely_typecast
 
 
 class NativeClient:
@@ -43,9 +43,10 @@ class NativeClient:
         url: str,
         screenshot: str | None,
     ) -> tuple[str, Data]:
-        return self.session.retriever_agent.invoke(
+        explanation, result = self.session.retriever_agent.invoke(
             statement, accessibility_tree, title=title, url=url, screenshot=screenshot
         )
+        return explanation, loosely_typecast(result)
 
     def find_area(self, description: str, accessibility_tree: str):
         return self.session.area_agent.invoke(description, accessibility_tree)
